Add  basic files for the landing page assignment
An image folger, short img was created to store the images of the webpage.
This was done by using the grafical user interface add a folder button.
A index.html file in combination with a style.css file were added.
This was achieved in the terminal using the touch command.